Moved VERSION out to file
