+++ LICENSE.txt
@@ -1,4 +1,6 @@
-Copyright (c) 2009-2012 Solutious Inc, Delano Mandelbaum
+The MIT License (MIT)
+
+Copyright (c) 2023 delano
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
